Split logic and controller code. Added env for test to prevent port conflicts and errors on requests due to keys.
