+++ src/routes/YouTubeVideoInfo.ts
@@ -1,35 +1,9 @@
-import { Router, Request, Response } from 'express'
-import YouTubeAxiosConfig from '../downstream-services/YouTubeAxiosConfig';
-import Constants from '../downstream-services/Constants';
-import Endpoint from "./Endpoint";
-import HeartAPIError from '../errors/HeartAPIError';
-import { AxiosError, AxiosResponse } from 'axios';
-import moize from 'moize';
-
-
-type YouTubeAPIResponse = {
-	kind: string,
-	etag: string,
-	items: YouTubeAPIResponseItem[],
-	pageInfo: {
-		totalResults: number,
-		resultsPerPage: number
-	}
-}
-
-type YouTubeAPIResponseItem =
-	{
-		kind: string,
-		etag: string,
-		id: string,
-		statistics: {
-			viewCount: string,
-			likeCount: string,
-			dislikeCount: string,
-			favoriteCount: string,
-			commentCount: string
-		}
-	}
+import { Router } from 'express'
+import YouTubeAxiosConfig from '../downstream-services/YouTubeAxiosConfig'
+import Endpoint from "./Endpoint"
+import { AxiosResponse } from 'axios'
+import moize from 'moize'
+import YouTubeVideoInfoController, { YouTubeAPIResponse, YouTubeAPIResponseItem } from '../controller/YouTubeVideoInfoController'
 
 
 type VideoInfoResponse = {
@@ -54,38 +28,12 @@ export default class YouTubeVideoInfo implements Endpoint {
 
 
 	get(): void {
-		this.router.get('/yt/video/info', this.getVideoInfoCallBack)
+		this.router.get('/yt/video/info', YouTubeVideoInfoController(this.memoizedYouTubeRequest, this.getVideoInfoResponse))
 	}
 
 
 	post(): void {
-		throw new Error("Method not implemented.");
-	}
-
-
-	private getVideoInfoCallBack = (req: Request, res: Response) => {
-		let status = 200
-		if (req.query == null || req.query.key == null || req.query.videoId == null) {
-			status = 400
-
-			res.status(status)
-			res.json(new HeartAPIError("Missing required query params.", status))
-			res.send()
-		} else if (req.query.key !== Constants.HEART_API_KEY) {
-			let status = 401
-
-			res.status(status)
-			res.json(new HeartAPIError("API key is incorrect.", status))
-			res.send()
-		} else {
-			this.memoizedYouTubeRequest(req.query.videoId as string)
-				.then((ytResponse: AxiosResponse<YouTubeAPIResponse>) => {
-					res.status(status)
-					res.json(this.getVideoInfoResponse(ytResponse.data))
-					res.send()
-				})
-				.catch((error: AxiosError) => YouTubeAxiosConfig.YOUTUBE_API_ERROR_CALLBACK(error, res))
-		}
+		throw new Error("Method not implemented.")
 	}
